feat(pickercomponent): convert to typescript
